feat: support state_agg and duration_in for state_agg
* bench: add state_agg_bench
