+++ query_server/sqllogicaltests/cases/function/state_agg/state_agg.slt
@@ -1,0 +1,14 @@
+statement ok
+alter database public set ttl '1000000d';
+
+statement ok
+create table if not exists states(state STRING);
+
+statement ok
+insert into states values
+('2020-01-01 10:00:00', 'starting'),
+('2020-01-01 10:30:00', 'running'),
+('2020-01-03 16:00:00', 'error'),
+('2020-01-03 18:30:00', 'starting'),
+('2020-01-03 19:30:00', 'running'),
+('2020-01-05 12:00:00', 'stopping');
